Add `attrs` setter shorthand, refactor
The `attrs` shorthand allows setting multiple fields in a single owbs.bind call. While this makes owbs a bit bigger, it allows for much denser binding code, so where owbs is in use it should pay for itself in bit size on average.

Also refactored and extracted some common elements and variables in order to allow them to transpile to single letters, meaning that reuse is more efficient.